# Reduce runtime of Go Encode() by another 25%