--- a/go/encode.go
+++ b/go/encode.go
@@ -54,7 +54,7 @@ func Encode(lat, lng float64, codeLen int) string {
 	}
 	// Use a char array so we can build it up from the end digits, without having
 	// to keep reallocating strings.
-	var code [16]byte
+	var code [maxCodeLen + 1]byte
 
 	// Compute the code.
 	// This approach converts each value to an integer after multiplying it by
@@ -64,18 +64,17 @@ func Encode(lat, lng float64, codeLen int) string {
 
 	// Compute the grid part of the code if necessary.
 	if codeLen > pairCodeLen {
-		code[sepPos+7], latVal, lngVal = latLngGrid(latVal, lngVal)
-		code[sepPos+6], latVal, lngVal = latLngGrid(latVal, lngVal)
-		code[sepPos+5], latVal, lngVal = latLngGrid(latVal, lngVal)
-		code[sepPos+4], latVal, lngVal = latLngGrid(latVal, lngVal)
-		code[sepPos+3], latVal, lngVal = latLngGrid(latVal, lngVal)
+		code[sepPos+7], latVal, lngVal = latLngGridStep(latVal, lngVal)
+		code[sepPos+6], latVal, lngVal = latLngGridStep(latVal, lngVal)
+		code[sepPos+5], latVal, lngVal = latLngGridStep(latVal, lngVal)
+		code[sepPos+4], latVal, lngVal = latLngGridStep(latVal, lngVal)
+		code[sepPos+3], latVal, lngVal = latLngGridStep(latVal, lngVal)
 	} else {
 		latVal /= gridLatFullValue
 		lngVal /= gridLngFullValue
 	}
 
-	// The first pair after the separator is a special case as it handles the
-	// transition from the grid.
+	// Add the pair after the separator.
 	latNdx := latVal % int64(encBase)
 	lngNdx := lngVal % int64(encBase)
 	code[sepPos+2] = Alphabet[lngNdx]
@@ -85,17 +84,18 @@ func Encode(lat, lng float64, codeLen int) string {
 	code[sepPos] = Separator
 
 	// Compute the pair section of the code.
-	code[7], lngVal = lngIter(lngVal)
-	code[6], latVal = latIter(latVal)
+	// Even indices contain latitude and odd contain longitude.
+	code[7], lngVal = pairIndexStep(lngVal)
+	code[6], latVal = pairIndexStep(latVal)
 
-	code[5], lngVal = lngIter(lngVal)
-	code[4], latVal = latIter(latVal)
+	code[5], lngVal = pairIndexStep(lngVal)
+	code[4], latVal = pairIndexStep(latVal)
 
-	code[3], lngVal = lngIter(lngVal)
-	code[2], latVal = latIter(latVal)
+	code[3], lngVal = pairIndexStep(lngVal)
+	code[2], latVal = pairIndexStep(latVal)
 
-	code[1], lngVal = lngIter(lngVal)
-	code[0], latVal = latIter(latVal)
+	code[1], lngVal = pairIndexStep(lngVal)
+	code[0], latVal = pairIndexStep(latVal)
 
 	// If we don't need to pad the code, return the requested section.
 	if codeLen >= sepPos {
@@ -105,6 +105,8 @@ func Encode(lat, lng float64, codeLen int) string {
 	return string(code[:codeLen]) + strings.Repeat(string(Padding), sepPos-codeLen) + string(Separator)
 }
 
+// roundLatLngToInts rounds the passed latitude and longitude to integral values
+// representing a location within 1 centimetre of the passed coordinates.
 func roundLatLngToInts(lat, lng float64) (int64, int64) {
 	// To round, we:
 	// 1) Offset latitude and longitude so that all values are positive.
@@ -118,13 +120,17 @@ func roundLatLngToInts(lat, lng float64) (int64, int64) {
 	latVal := int64((lat+latMax)*finalLatPrecision*(1<<roundPrecision)+1) >> roundPrecision
 	lngVal := int64((lng+lngMax)*finalLngPrecision*(1<<roundPrecision)+1) >> roundPrecision
 	return latVal, lngVal
-
 }
 
+// clipCodeLen returns the smallest valid code length greater than or equal to
+// the desired code length.
 func clipCodeLen(codeLen int) int {
 	if codeLen <= 0 {
+		// Default to a full pair code if codeLen is the default or negative
+		// value.
 		return pairCodeLen
 	} else if codeLen < pairCodeLen && codeLen%2 == 1 {
+		// Codes only consisting of pairs must have an even length.
 		return codeLen + 1
 	} else if codeLen > maxCodeLen {
 		return maxCodeLen
@@ -132,23 +138,22 @@ func clipCodeLen(codeLen int) int {
 	return codeLen
 }
 
-func latLngGrid(latVal, lngVal int64) (byte, int64, int64) {
+// latLngGridStep computes the next smallest grid code in sequence,
+// followed by the remaining latitude and longitude values not yet converted
+// to a grid code.
+func latLngGridStep(latVal, lngVal int64) (byte, int64, int64) {
 	latDigit := latVal % int64(gridRows)
 	lngDigit := lngVal % int64(gridCols)
 	ndx := latDigit*gridCols + lngDigit
 	return Alphabet[ndx], latVal / int64(gridRows), lngVal / int64(gridCols)
 }
 
-func latIter(latVal int64) (byte, int64) {
-	latVal /= int64(encBase)
-	latNdx := latVal % int64(encBase)
-	return Alphabet[latNdx], latVal
-}
-
-func lngIter(lngVal int64) (byte, int64) {
-	lngVal /= int64(encBase)
-	lngNdx := lngVal % int64(encBase)
-	return Alphabet[lngNdx], lngVal
+// pairIndexStep computes the next smallest pair code in sequence,
+// followed by the remaining integer not yet converted to a pair code.
+func pairIndexStep(coordinate int64) (byte, int64) {
+	coordinate /= int64(encBase)
+	latNdx := coordinate % int64(encBase)
+	return Alphabet[latNdx], coordinate
 }
 
 // computeLatPrec computes the precision value for a given code length.