--- a/python/openlocationcode/__init__.py
+++ b/python/openlocationcode/__init__.py
@@ -1,1 +1,4 @@
-from openlocationcode import *
+if __name__ == "__main__":
+    from openlocationcode import *
+else:
+    from .openlocationcode import *