This is (another) performance optimization change for the Go implementation of Encode(), which reduces the runtime by about 25%. This is the additional performance improvement I mentioned in #566.

There are two main changes here to accomplish this, either of which may be controversial as they do harm the code's readability. I have attempted to mitigate this with refactorings, and verified the extracted functions are inlined to avoid the additional overhead of function calls.

The first change is that all loops are unrolled. Given that all loops are unconditionally executed a constant number of times, this can be done without introducing any additional branches to the code.

The second change is that the precision rounding logic is modified to require zero divisions. This new method is significantly faster, but may change the final digits of codes requiring sub-centimeter precision (I don't believe this library has this as a supported use case?). There are no cases in the test suite in which this causes a difference, but if desired I can likely find such an edge case.

Before:

```
will@Janeway:~/GolandProjects/open-location-code/go$ go test --run=NONE --bench=Encode --benchtime=10s -cpuprofile=/home/will/cpu2.out
goos: linux
goarch: amd64
pkg: github.com/google/open-location-code/go
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor 
BenchmarkEncode-64    	186547320	        64.70 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/google/open-location-code/go	22.222s
```

After:

```
will@Janeway:~/GolandProjects/open-location-code/go$ git checkout -
Switched to branch 'optimize-2'
will@Janeway:~/GolandProjects/open-location-code/go$ go test --run=NONE --bench=Encode --benchtime=10s -cpuprofile=/home/will/cpu.out
goos: linux
goarch: amd64
pkg: github.com/google/open-location-code/go
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor 
BenchmarkEncode-64    	233725107	        49.01 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/google/open-location-code/go	21.471s
```

CPU profile of before:
![profile002](https://github.com/user-attachments/assets/a6238db7-84f5-42ef-8478-84ed916d01b6)


CPU profile of after:
![profile001](https://github.com/user-attachments/assets/837c2c06-d508-4d72-bbac-6a8d6fa1a3e4)