@willbeason 2025-01-06
Okay, improved comments and renamed functions as requested.